Frewtzapi/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tokens==0.0.3
tzdata==2025.2
